requirements.txt
# Core Deep Learning
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Data Processing
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
opencv-python==4.8.0.74

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
tensorboard==2.13.0

# Progress Bars & Logging
tqdm==4.65.0

# Scientific Computing
scikit-learn==1.3.0
scipy==1.11.1

# Image Augmentation
albumentations==1.3.1

# Path Planning (not strictly needed but nice to have)
# networkx==3.1

# Utilities
pyyaml==6.0.1
# wget==3.2  # Not needed on Windows

# Jupyter (for Colab notebooks)
# jupyter==1.0.0  # Install separately if needed
# ipywidgets==8.1.0
